+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next
 
+* Update to new `Cursive` `API` - Breaking change
+    * Replace all `Box<AnyView>` with `ViewBox`
 * Improved selecting first item in `Autocomplete`
 * Fix scrolling view `Autocomplete` when list empty
 * Improved focus change between `edit` and `select` in `Autocomplete`